Rewrite pkgfile in python
The purpose of this conversion is to use a database file more efficient
than the database based on a tree of files, to provide faster access
times and a better user experience.

For example, "time pkgfile -l xscreensaver"
gives for the bash version
real	0m28.280s
user	0m0.597s
sys	0m0.647s

whereas the new one gives:
real	0m2.391s
user	0m0.220s
sys	0m0.173s

with an empty OS filesystem cache each time
Cache dropped with:
echo 3 > /proc/sys/vm/drop_caches
echo 0 > /proc/sys/vm/drop_caches

About sqlite choice:
At first, I wanted to use bsddb but bsddb module will be dropped in python
3.x. So it will need pybsddb, so this makes another dependancy.
hamsterdb is another possibility; it's GPL.
But sqlite3 is in the standard library of python, so I go for it.

The downloaded .files.tar.gz are converted to a sqlite file when
updating the file lists.
During the update, the local pacman database is also converted to a sqlite
db file:
	- to speed up access to pacman db ;-)
	- to be consistent with other dbs
	- to avoid using a pacman process
Note: However pacman is called (at least twice in the code) to get
pacman db path and to get repo and mirrors

Databases created from files.tar.gz are created from strach every time.
This could change in the future.
The local pacman db is only first created from strach (this need at lot
of I/O) but to avoid that, the next call to -u/--update will just update
the pkg in local.db file.

With this conversion, using CMD_SEARCH_ENABLED=1 would, at last, be
something usefull and not painfull to use. You will not have time
to go and grab some coffee :-P

Also, it is now possible to use --info for sync database, something that
was not done in the original bash script.

However, use of wget has been dropped to use python standard library
(urllib.urlretrieve) so RATELIMIT is unsupported. This could be changed
but simply calling an external process for almost nothing.
I just spotted bug #20697 http://bugs.archlinux.org/task/20697 that
noone seems to have noticed so far. So I guess RATELIMIT is not a very
used option.

The database:
The format of the sqlite database is simply a row for each package.
* lists are pickled and zlib compressed (mainly for files list)
It is posible to avoid these pythonistic things to be able to read those
dbs from C for example. This will need some adaptation like a simple
marshalling of lists (of strings); concatenate them for example.

convert_sqlite.py could be used as a standalone script to convert
.files.tar.gz or a tree like local pacman db into a sqlite file.
Converting extra.files.tar.gz takes around 40s here with 20MB of memory
at most.

Original authors will easily go around the pkgfile source code because
it retains the same structure and functions.

TODO:
* finish the implementation (search is not done yet)
* either paste convert_sqlite.py into pkgfile or find a way to
import it when installed system wide
* find a better name for convert_sqlite.py
* polish it and kill bugs
* name it pkgfile or pkgfile.py ?

I just notice I forgot the -r/--regex action.
It's almost the same as -g so I do not plan to implement it (yet ?)